[Update] 1-1, 1-2, 1-3 스테이지수정 -SG
맵 크기 줄임
게임오버 / 죽는 에니메이션 다시만듬
필요없는 스크립트 삭제
