+++ Assets/Animation/GAMEOVER/Boooooooom.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b7ff11f8e4975141aa19c252541167a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
