+++ Assets/Scenes/Stage1-1.unity
@@ -446,7 +446,7 @@ MonoBehaviour:
   player: {fileID: 904379937}
   smoothing: 0.2
   minCameraBoundary: {x: 0, y: 0}
-  maxCameraBoundary: {x: 1000, y: 10}
+  maxCameraBoundary: {x: 105, y: 15}
 --- !u!81 &741140524
 AudioListener:
   m_ObjectHideFlags: 0
@@ -490,7 +490,7 @@ Camera:
   far clip plane: 1000
   field of view: 60
   orthographic: 1
-  orthographic size: 17.1
+  orthographic size: 10
   m_Depth: -1
   m_CullingMask:
     serializedVersion: 2
@@ -1005,7 +1005,7 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 2.56, y: 2.56}
+  m_Size: {x: 1.7, y: 1.7}
   m_EdgeRadius: 0
 --- !u!114 &904379934
 MonoBehaviour:
@@ -1024,6 +1024,7 @@ MonoBehaviour:
   SpriteRight: {fileID: 21300028, guid: 15b407f3e228b484c8a8b51a211b559b, type: 3}
   movePower: 5
   jumpPower: 10
+  maxHealth: 1
 --- !u!50 &904379935
 Rigidbody2D:
   serializedVersion: 4
@@ -1111,7 +1112,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 904379932}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -27.05, y: -15.66, z: 0}
+  m_LocalPosition: {x: -16, y: -7.8, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1651,7 +1652,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1486185281}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -0.2, y: -2.88, z: 0}
+  m_LocalPosition: {x: 11.39, y: 5.16, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -7831,7 +7832,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1939930532}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 106.4, y: -15.5, z: 0}
+  m_LocalPosition: {x: 118.69, y: -7.08, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
