+++ Assets/Scenes/Stage2-3.unity
@@ -757,7 +757,7 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 2.56, y: 2.56}
+  m_Size: {x: 1.7, y: 1.7}
   m_EdgeRadius: 0
 --- !u!114 &73014265
 MonoBehaviour:
@@ -776,6 +776,7 @@ MonoBehaviour:
   SpriteRight: {fileID: 21300028, guid: 15b407f3e228b484c8a8b51a211b559b, type: 3}
   movePower: 5
   jumpPower: 10
+  maxHealth: 1
 --- !u!50 &73014266
 Rigidbody2D:
   serializedVersion: 4
@@ -3291,7 +3292,7 @@ Camera:
   far clip plane: 1000
   field of view: 60
   orthographic: 1
-  orthographic size: 8.510744
+  orthographic size: 10
   m_Depth: -1
   m_CullingMask:
     serializedVersion: 2
@@ -3346,7 +3347,7 @@ MonoBehaviour:
   player: {fileID: 73014268}
   smoothing: 0.2
   minCameraBoundary: {x: 0, y: 0}
-  maxCameraBoundary: {x: 2000, y: 40}
+  maxCameraBoundary: {x: 105, y: 15}
 --- !u!1 &369325679
 GameObject:
   m_ObjectHideFlags: 0
@@ -19210,9 +19211,10 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 809999017}
+  - component: {fileID: 809999018}
   m_Layer: 0
   m_Name: Circle
-  m_TagString: Untagged
+  m_TagString: obstacle
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -19272,6 +19274,18 @@ Transform:
   m_Father: {fileID: 769775418}
   m_RootOrder: -1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &809999018
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 809999016}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 03dbfec574dec974b9d03bfaa939dd8a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &826232152
 GameObject:
   m_ObjectHideFlags: 0
