+++ Assets/Scenes/TestScenes/GameOver.unity
@@ -215,6 +215,32 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &1439700531 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 8774130827924538856, guid: 4f7e97443ffd84bb79f5879acd8eda40, type: 3}
+  m_PrefabInstance: {fileID: 2939891398648328396}
+  m_PrefabAsset: {fileID: 0}
+--- !u!95 &1439700533
+Animator:
+  serializedVersion: 5
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1439700531}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: 7b7ff11f8e4975141aa19c252541167a, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
 --- !u!1001 &2939891398648328396
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -339,8 +365,12 @@ PrefabInstance:
       propertyPath: m_Controller
       value: 
       objectReference: {fileID: 9100000, guid: fe0f3a8f6264eb843b3203b16cfd832a, type: 2}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 347083742706785302, guid: 4f7e97443ffd84bb79f5879acd8eda40, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 8774130827924538856, guid: 4f7e97443ffd84bb79f5879acd8eda40, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1439700533}
   m_SourcePrefab: {fileID: 100100000, guid: 4f7e97443ffd84bb79f5879acd8eda40, type: 3}
