--- Assets/Scripts/Player/player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6cf48afc48df8444faaa75ede8be329a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
