--- Assets/Scripts/obstacle/CloudController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f62e3802903472449bee81848ebb07ae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
